[IMP] first steps with new account_check
