+++ addons/account_check/models/__init__.py
@@ -3,8 +3,5 @@
 # directory
 ##############################################################################
 from . import account_move
-from . import account_checkbook
-from . import account_check_operation
-from . import account_check
 from . import account_payment
 from . import account_chart_template
