+++ addons/account_check/models/account_chart_template.py
@@ -1,7 +1,3 @@
-##############################################################################
-# For copyright and license notices, see __manifest__.py file in module root
-# directory
-##############################################################################
 from odoo import models, _
 import logging
 _logger = logging.getLogger(__name__)
@@ -11,13 +7,8 @@ class AccountChartTemplate(models.Model):
     _inherit = 'account.chart.template'
 
     def _create_bank_journals(self, company, acc_template_ref):
-        """
-        Creates third checks journal and enable own checks on banks
-        """
-
         res = super(AccountChartTemplate, self)._create_bank_journals(company, acc_template_ref)
 
-        # creamos diario para cheques de terceros
         self.env['account.journal'].create({
             'name': _('Third Checks'),
             'type': 'cash',
@@ -29,7 +20,6 @@ class AccountChartTemplate(models.Model):
                 (4, self.env.ref('account_check.account_payment_method_out_third_checks').id, None),
             ],
         })
-        # creamos diario para cheques de terceros rechazados
         self.env['account.journal'].create({
             'name': _('Rejected Third Checks'),
             'type': 'cash',
