+++ addons/account_check/models/account_check_operation (borrar).py
