--- addons/account_check/security/account_check_security.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data noupdate="1">
-        <record id="account_check_rule" model="ir.rule">
-            <field name="name">Check Multi-Company</field>
-            <field name="model_id" ref="model_account_check"/>
-            <field eval="True" name="global"/>
-            <field name="domain_force">['|',('company_id', 'in', company_ids),('company_id','=',False)]</field>
-        </record>
-
-        <record id="account_check_operation_rule" model="ir.rule">
-            <field name="name">Check Operation Multi-Company</field>
-            <field name="model_id" ref="model_account_check_operation"/>
-            <field eval="True" name="global"/>
-            <field name="domain_force">['|',('check_id.company_id', 'in', company_ids),('check_id.company_id','=',False)]</field>
-        </record>
-    </data>
-</odoo>
