--- addons/account_check/security/ir.model.access.csv
@@ -1,5 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-account_check_access_full,account_check_access_full,model_account_check,account.group_account_invoice,1,1,1,1
-account_check_access_global,account_check_access_global,model_account_check,base.group_user,1,0,0,0
-account_check_operation_access_global,account_check_operation_access_global,model_account_check_operation,,1,0,0,0
-account_check_operation_access_full,account_check_operation_access_full,model_account_check_operation,account.group_account_invoice,1,1,1,1
