--- addons/account_check/views/account_chart_template_view.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="view_account_chart_template_form" model="ir.ui.view">
-        <field name="name">account.chart.template.form</field>
-        <field name="model">account.chart.template</field>
-        <field name="inherit_id" ref="account.view_account_chart_template_form"/>
-        <field name="arch" type="xml">
-            <field name="property_account_income_id" position="after">
-                <field name="rejected_check_account_id"/>
-                <field name="deferred_check_account_id"/>
-                <field name="holding_check_account_id"/>
-            </field>
-        </field>
-    </record>
-</odoo>
