+++ addons/account_check/wizards/account_payment_register_views.xml
@@ -8,34 +8,33 @@
         <field name="arch" type="xml">
             <form>
                 <field name="available_check_ids" invisible="1"/>
-                <field name="check_type" invisible="1"/>
-                <field name="payment_method_code" invisible="1"/>
             </form>
             <xpath expr="//div[@name='amount_div']//field[@name='amount']" position="attributes">
-                <attribute name="attrs">{'readonly': [('check_type', '!=', False)]}</attribute>
+                <attribute name="attrs">{'readonly': [('payment_method_code', 'in', ['in_third_checks', 'out_third_checks'])]}</attribute>
                 <attribute name="force_save">1</attribute>
             </xpath>
             <field name="payment_method_id" position="after">
-                <label for="check_id" attrs="{'invisible': [('check_type', '=', False)]}"/>
-                <div class="o_row" attrs="{'invisible': ['|', ('check_type', '=', False), '|', ('can_edit_wizard', '=', False), '&amp;', ('can_group_payments', '=', True), ('group_payment', '=', False)]}">
-                    <!-- select check -->
+
+                <!-- <div class="o_row" attrs="{'invisible': ['|', ('check_type', '=', False), '|', ('can_edit_wizard', '=', False), '&amp;', ('can_group_payments', '=', True), ('group_payment', '=', False)]}"> -->
                     <field
                         name="check_id"
-                        attrs="{'invisible': [('payment_method_code', 'not in', ['in_own_checks', 'in_third_checks', 'out_third_checks'])], 'required': [('payment_method_code', 'in', ['in_own_checks', 'in_third_checks', 'out_third_checks'])]}"
+                        attrs="{'invisible': [('payment_method_code', 'not in', ['in_third_checks', 'out_third_checks'])], 'required': [('payment_method_code', 'in', ['in_third_checks', 'out_third_checks'])]}"
                         domain="[('id', 'in', available_check_ids)]"
                         options="{'no_create': True}"
                     />
-                    <!-- new check -->
-                    <field
-                        name="check_id"
-                        attrs="{'invisible': [('payment_method_code', 'not in', ['new_third_checks', 'new_own_checks'])], 'required': [('payment_method_code', 'in', ['new_third_checks', 'new_own_checks'])]}"
-                        domain="[('id', 'in', available_check_ids)]"
-                        context="{'default_amount': amount, 'default_company_id': company_id, 'default_first_partner_id': partner_id, 'default_journal_id': journal_id, 'default_currency_id': currency_id, 'default_type': check_type, 'from_payment': True}"
-                    />
-                </div>
-                <div colspan="2" class="o_row" attrs="{'invisible': ['|', ('check_type', '=', False), '&amp;', ('can_edit_wizard', '=', True), '|', ('can_group_payments', '=', False), '&amp;', ('can_group_payments', '=', True), ('group_payment', '=', True)]}">
+                    <field name="check_number" readonly="0" attrs="{'invisible': [('payment_method_code', '!=', 'new_third_checks')], 'required': [('payment_method_code', '=', 'new_third_checks')]}"/>
+                    <field name="check_payment_date" attrs="{'invisible': [('payment_method_code', '!=', 'new_third_checks')]}"/>
+                    <field name="third_check_issue_date" string="Check Issue Date" attrs="{'invisible': [('payment_method_code', '!=', 'new_third_checks')], 'required': [('payment_method_code', '=', 'new_third_checks')]}"/>
+                    <field name="third_check_bank_id" string="Check Bank" attrs="{'invisible': [('payment_method_code', '!=', 'new_third_checks')], 'required': [('payment_method_code', '=', 'new_third_checks')]}"/>
+                    <field name="third_check_issuer_vat" string="Check Issuer Vat" attrs="{'invisible': [('payment_method_code', '!=', 'new_third_checks')], 'required': [('payment_method_code', '=', 'new_third_checks')]}"/>
+                    <field name="third_check_issuer_name" string="Check Issuer" attrs="{'invisible': [('payment_method_code', '!=', 'new_third_checks')], 'required': [('payment_method_code', '=', 'new_third_checks')]}"/>
+                <!-- </div> -->
+
+                <!-- <div colspan="2" class="o_row" attrs="{'invisible': ['|', ('check_type', '=', False), '&amp;', ('can_edit_wizard', '=', True), '|', ('can_group_payments', '=', False), '&amp;', ('can_group_payments', '=', True), ('group_payment', '=', True)]}">
                     <p class="alert alert-warning" role="alert">No puede usar cheques en grupos de pagos de distintos partners o sin agrupar!</p>
-                </div>
+                </div> -->
+
+
             </field>
         </field>
     </record>
