+++ addons/account_check_printing/__manifest__.py
@@ -21,7 +21,8 @@ The check settings are located in the accounting journals configuration page.
         'views/account_payment_views.xml',
         'views/res_config_settings_views.xml',
         'views/res_partner_views.xml',
-        'wizard/print_prenumbered_checks_views.xml'
+        'wizard/account_payment_register_views.xml',
+        'wizard/print_prenumbered_checks_views.xml',
     ],
     'installable': True,
     'auto_install': False,
