+++ addons/account_check_printing/wizard/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import account_payment_register
 from . import print_prenumbered_checks
